Use proper GeoJSON types and fix display of Regions (#90)
* Use proper GeoJSON types and fix display of Regions

* Delete package-lock

* Zoom to footprint when toggling on

* Fix package resolution of .ts files

* Fix selection of geometry input

* Add // eslint-disable-line for geojson import

* Fix search button

* Move useMap

* HACK

* Code cleanliness

* Fix search error

* Fix GeoJSON form type checking

* Fix adding search geometry to map

* Remove geojsonhint

* Use GeoJSON source for footprints

* Make cesium store a submodule

* Refactor cesium store

* Major refactoring of store
