+++ package.json
@@ -12,11 +12,11 @@
     "@girder/oauth-client": "^0.7.6",
     "@open-wc/webpack-import-meta-loader": "^0.4.7",
     "@sentry/vue": "^6.2.0",
+    "@types/geojson": "^7946.0.8",
     "@vue/composition-api": "^1.0.0-rc.2",
     "axios": "^0.21.1",
     "cesium": "^1.85.0",
     "copy-webpack-plugin": "^6.4.1",
-    "geojsonhint": "^2.0.0",
     "strip-pragma-loader": "^1.0.0",
     "vue": "^2.6.11",
     "vue-router": "^3.2.0",
