+++ src/components/molecules/GeoJsonForm.vue
@@ -1,9 +1,9 @@
 <script lang="ts">
 import { defineComponent, ref } from '@vue/composition-api';
 import {
-  useMap, drawnShape, specifiedShape, geoJsonShape,
-} from '@/store';
-import { hint } from 'geojsonhint';
+  drawnShape, specifiedShape, geometryInputSelection,
+} from '@/store/search';
+import { useMap } from '@/store/cesium/search';
 
 export default defineComponent({
   name: 'GeoJsonForm',
@@ -19,19 +19,23 @@ export default defineComponent({
 
     const clearShape = () => {
       drawnShape.value = {
-        type: '',
+        type: 'Polygon',
         coordinates: [],
       };
       geoJsonErrorMessages.value = [];
     };
     const selectShape = (value: string) => {
       clearShape();
-      geoJsonShape.value = value;
+      geometryInputSelection.value = value;
     };
 
     const isGeoJSON = (inputText: string) => {
-      const validation = hint(inputText);
-      geoJsonErrorMessages.value = validation.map((error: {message: string}) => error.message);
+      try {
+        JSON.parse(inputText);
+        geoJsonErrorMessages.value = [];
+      } catch (e) {
+        geoJsonErrorMessages.value = ['Not a valid GeoJSON Polygon or MultiPolygon'];
+      }
       return true;
     };
     const confirmGeoJSON = () => {
@@ -62,13 +66,13 @@ export default defineComponent({
       };
       return true;
     };
-
     return {
       useMap,
       clearShape,
       selectShape,
       drawnShape,
-      geoJsonShape,
+      specifiedShape,
+      geometryInputSelection,
       geoOptions,
       geoJsonString,
       geoJsonErrorMessages,
@@ -84,11 +88,11 @@ export default defineComponent({
 <template>
   <div>
     <v-select
-      v-model="geoJsonShape"
+      v-model="geometryInputSelection"
       :items="geoOptions"
       label="Search area"
       messages="Select a method of specifying a geographical area."
-      :hide-details="geoJsonShape !== undefined"
+      :hide-details="geometryInputSelection !== undefined"
       outlined
       clearable
       dense
@@ -96,7 +100,7 @@ export default defineComponent({
       @change="selectShape"
     />
     <v-btn
-      v-if="geoJsonShape === geoOptions[0]"
+      v-if="geometryInputSelection === geoOptions[0] && !useMap"
       color="#188DC8"
       block
       class="mt-3"
@@ -104,12 +108,12 @@ export default defineComponent({
     >
       Draw polygon on map
     </v-btn>
-    <div v-if="geoJsonShape === geoOptions[0] && useMap">
+    <div v-if="geometryInputSelection === geoOptions[0] && useMap">
       Click on the map to draw points of a polygon.
       Double click to complete the polygon selection.
     </div>
     <div
-      v-if="geoJsonShape === geoOptions[1]"
+      v-if="geometryInputSelection === geoOptions[1]"
     >
       Paste GeoJSON contents below.
       <v-spacer />
@@ -130,7 +134,7 @@ export default defineComponent({
         Confirm GeoJSON search area
       </v-btn>
     </div>
-    <div v-if="geoJsonShape === geoOptions[2]">
+    <div v-if="geometryInputSelection === geoOptions[2]">
       <v-file-input
         accept=".json,.txt"
         autofocus
