--- src/shims-geojsonhint.d.ts
@@ -1,5 +1,0 @@
-declare module 'geojsonhint' {
-  import geojsonhint from 'geojsonhint';
-
-  export default geojsonhint;
-}
