--- src/types/geojson.d.ts
@@ -1,6 +1,0 @@
-declare module 'geojsonhint' {
-  function hint(input: string): Array<{
-    message: string;
-    line: number;
-  }>;
-}
